Fixes ICNT86X touch controller support
Fixes support for the ICNT86X touch controller, including initialization, coordinate scanning, and transformation.

Refactors display and touch type definitions into a shared header for better code organization and reusability.

Implements coordinate transformation for GT1151 and ICNT86X touch controllers to handle different display mirroring modes.

+++ main/touch/TouchPoint.h
@@ -13,9 +13,9 @@ public:
     TouchPoint();
     ~TouchPoint();
     uint8_t track_id; // the trackId of the touch point
-    uint8_t x;        // x-axis value/coordinate
-    uint8_t y;        // y-axis value/coordinate
-    uint8_t size;     // size of the touch point
+    uint16_t x;       // x-axis value/coordinate (16-bit from controller)
+    uint16_t y;       // y-axis value/coordinate (16-bit from controller)
+    uint16_t size;    // size of the touch point (sometimes 16-bit)
 };
 
 #endif
